refactor(test): P5-1 follow-up — extract SVG widget JS bridge cluster
Continue test_widget_bridge.cpp split. Extracts the three SVG-widget
JS bridge clusters (802 lines combined) into a sibling TU:

  1. pulp #965 — SvgPathWidget JS bridge integration (lines 2800-3029).
  2. pulp #1416 — SvgRectWidget + SvgLineWidget JS bridge integration
     (lines 3794-4311).
  3. Compound-path SVG icons (Spectr PEAK / AVG / BOTH / OFF) —
     regression-pins the parser's enumeration of every M and L across
     disjoint subpaths (lines 10948-11001).

* test_widget_bridge.cpp: 11,502 → 10,700 (-802)
* New target pulp-test-widget-bridge-svg: 850 lines, 21 TEST_CASEs,
  156 assertions.

Non-contiguous source ranges consolidated into one TU. canvasFromBridge
duplicated locally per the existing extracted-TU pattern.

This is commit 1 of 2 in a bundled-cuts PR (HTML ARIA + querySelector
follows; single version bump at the end).
