+++ test/CMakeLists.txt
@@ -1781,6 +1781,12 @@ pulp_add_test_suite(pulp-test-widget-bridge-yoga LIBRARIES pulp::view)
 # compat.json entries.
 pulp_add_test_suite(pulp-test-widget-bridge-canvas2d-wave2 LIBRARIES pulp::view)
 
+# Widget bridge — SVG widgets (P5-1 follow-up). Three clusters:
+# pulp #965 SvgPathWidget JS bridge integration + pulp #1416
+# SvgRectWidget + SvgLineWidget JS bridge integration + compound-path
+# parser regression (Spectr PEAK / AVG / BOTH / OFF analyzer icons).
+pulp_add_test_suite(pulp-test-widget-bridge-svg LIBRARIES pulp::view)
+
 # Autonomous Spectr regression suite (pulp-internal task #63, follow-up to
 # pulp #1792). Composes the six [contract] invariants from
 # test_widget_bridge.cpp into a Spectr-shaped mini-scenario so a future
